automate/requirements.txt
pandas
python-dotenv
deta